Parse Universal Transaction flags from TxFlags.h, extract shared rippled-source-reading helpers, and rename sflags.rs to tx_flags.rs (updating generate-sflags.sh's output default to match)
